Avoid using empty expression + applyTagFilter moved from Event to ReportBuilder
Because it uses Connection so it makes more sense to have it in the builder
